put getDelegates function into a dedicated services file
