+++ src/app/api/delegates/route.ts
@@ -1,23 +1,7 @@
 import { NextResponse } from "next/server";
-import { fetchDuneData } from "@/services/dune";
-import type { FormattedDelegate } from "@/types/serverTypes";
+import { getDelegates } from "@/services/getDelegates";
 
 export async function GET() {
 	const delegates = await getDelegates();
 	return NextResponse.json(delegates)
-}
-
-async function getDelegates() {
-	const duneData = await fetchDuneData()
-	if (!duneData) return
-	const formattedDelegateData = duneData.map((delegate) => {
-		return {
-			rank: delegate.delegate_rank,
-			address: delegate.delegate,
-			username: delegate.delegate_name_raw,
-			voting_power: delegate.dt_voting_power,
-			pct_voting_power: delegate.pct_voting_power
-		} as FormattedDelegate
-	})
-	return formattedDelegateData
 }
